works, with extra logging
This code registers a custom topic in the dropdown menu on the webhooks page in the backend. You can select it, and modify the code in this plugin, so it does what you want. The business logic logs are in the native woocommerce logs (currently woocommerce - status - logs) and the default webhook delivery logs are also available, perhaps you have to turn them on.