works and changes status
after the webhook is sent, change the status to fulfillio. the status is created in functions.php. even though i dont want the option manually change the status to fulfillio, the code didnt work otherwise, and it always went to payment processing status. not sure why it has to be present in the ui to work, but this way it does