works, some things are todo
orders are now automatically completed. we also have the potential of alerting when a fulfillio order has not been successfully sent to the fulfillio api. we can also send the tracking number to the user. and we can also add send the tracking to paypal in case he paid with paypal. and we can also check when the package was sent, and if a lot of time passes, make an alert.