Changed plugin description and daily task
It often happens that a user would like to disable the sending of orders to the external system, for example, if stock in the fulfillment center is low. I have tested it and disabling the webhook works as expected - the order is not sent externally. Disabling the plugin to get the same functionality is not necessary, because we lose the automatic order completion and status changes that way.